refactor: Update DiscordRichPresence struct to support multiple buttons
The DiscordRichPresence struct in discord_rpc.h has been modified to include support for multiple buttons. This change introduces a new struct called DiscordButton, which contains the label and URL for each button. The buttons array in DiscordRichPresence can now hold up to 2 buttons.

This refactor allows for more interactive and engaging presence updates in Discord applications.

It is not yet etablished in the write process, so using buttons will currently break the programm.