refactor: Update Makefile to generate object files in a separate directory
This commit modifies the Makefile to generate object files in a separate "bin" directory. Previously, the object files were generated in the same directory as the source files, which cluttered the source directory. Now, the object files are generated in the "bin" directory, improving the organization of the project.

Also tested the discord rpc buttons feature.